edited bun class
